+++ R/metadata.R
@@ -32,11 +32,9 @@
 #'
 #' xportr_metadata(adlb, metadata, "test")
 #'
-#' library(magrittr)
-#'
-#' adlb %>%
-#'   xportr_metadata(metadata, "test") %>%
-#'   xportr_type() %>%
+#' adlb |>
+#'   xportr_metadata(metadata, "test") |>
+#'   xportr_type() |>
 #'   xportr_order()
 xportr_metadata <- function(.df,
                             metadata = NULL,