+++ R/xportr-package.R
@@ -111,17 +111,13 @@
 #'   %||%
 #' @importFrom dplyr left_join bind_cols filter select rename rename_with n
 #'   everything arrange group_by summarize mutate ungroup case_when distinct
-#'   tribble if_else across as_tibble bind_rows
+#'   tribble if_else across as_tibble bind_rows all_of any_of where
 #' @importFrom glue glue glue_collapse
 #' @importFrom cli cli_alert_info cli_h2 cli_alert_success cli_div cli_text
 #'   cli_alert_danger cli_warn start_app stop_app
-#' @importFrom tidyselect all_of any_of where
 #' @importFrom utils capture.output str tail packageVersion
-#' @importFrom stringr str_detect str_extract str_replace str_replace_all str_sub
-#' @importFrom readr parse_number
 #' @importFrom purrr map_chr map2_chr walk iwalk map map_dbl pluck map_lgl
 #' @importFrom graphics stem
-#' @importFrom magrittr %>% extract2
 #' @importFrom checkmate assert assert_character assert_choice assert_data_frame assert_subset
 #' assert_integer assert_logical assert_numeric assert_string makeAssertion check_data_frame
 #' check_r6 test_data_frame test_string vname